Split models catalog helper logic
